# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/acpc/mobilepatienttracker/ConsultationDetails.java
#	app/src/main/res/values/styles.xml

+++ app/src/main/java/com/acpc/mobilepatienttracker/DBookingDetails.java
@@ -1,7 +1,6 @@
 package com.acpc.mobilepatienttracker;
 
 import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -22,26 +21,14 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
-import com.google.firestore.v1.WriteResult;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-
-import static android.widget.Toast.LENGTH_LONG;
-import static android.widget.Toast.makeText;
-
-public class
-DBookingDetails extends AppCompatActivity {
+public class DBookingDetails extends AppCompatActivity {
 
     String name;
     String id;
@@ -59,9 +46,6 @@ DBookingDetails extends AppCompatActivity {
     private Button accept;
     private Button reject;
 
-
-    FirebaseAuth mAuth;
-    final FirebaseFirestore database = FirebaseFirestore.getInstance();
 
     private Doctor doc = new Doctor();
 
@@ -92,10 +76,27 @@ DBookingDetails extends AppCompatActivity {
 
         }
 
+        pullBookingData(id);
+
         nameT.setText(name);
         idT.setText(id);
         dateT.setText(date);
         timeT.setText(time);
+
+
+    }
+
+
+    public void pullBookingData(String test)
+    {
+        if(test.equals("___NULL_DEV___"))
+        {
+            addPatient("___NULL_DEV___");
+            DeleteBooking("___NULL_DEV___");
+            return;
+
+        }
+        final FirebaseFirestore database = FirebaseFirestore.getInstance();
 
         final FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
 
@@ -109,7 +110,7 @@ DBookingDetails extends AppCompatActivity {
                         doc_id = ID;
                         // Toast.makeText(DBookingDetails.this, "Hi", Toast.LENGTH_SHORT).show();
 
-                       // final Bookings booking = new Bookings(name, id, date, time, doc_id);
+                        // final Bookings booking = new Bookings(name, id, date, time, doc_id);
                         //final AcceptReject s = new AcceptReject();
 
 
@@ -117,12 +118,12 @@ DBookingDetails extends AppCompatActivity {
                             @Override
                             public void onClick(View v) {
 
-                                addPatient();
+                                addPatient("1");
 
-                                SplashActivity.msgserv.sendNotification("Booking accepted for patient: "+name);
+//                                SplashActivity.msgserv.sendNotification("Booking accepted for patient: "+name);
 
                                 aOrR = "Accepted";
-                                final AccOrRej s = new AccOrRej(name,id,date,time,doc_id, aOrR);
+                                final AccOrRej s = new AccOrRej(name,id,date,time,doc_id, aOrR,"Incomplete");
 
                                 //      Toast.makeText(DBookingDetails.this, "Processing booking", LENGTH_LONG).show();
 
@@ -162,9 +163,9 @@ DBookingDetails extends AppCompatActivity {
                             public void onClick(View v) {
 
                                 //  Toast.makeText(DBookingDetails.this, "Declining booking", LENGTH_LONG).show();
-                                SplashActivity.msgserv.sendNotification("Booking rejected for patient: "+name);
+//                                SplashActivity.msgserv.sendNotification("Booking rejected for patient: "+name);
                                 aOrR = "Rejected";
-                                final AccOrRej s = new AccOrRej(name,id,date,time,doc_id, aOrR);
+                                final AccOrRej s = new AccOrRej(name,id,date,time,doc_id, aOrR,"");
 
                                 //  UpdateDPatientList(doc_id, id);
 
@@ -206,12 +207,16 @@ DBookingDetails extends AppCompatActivity {
 
             }
         });
-
-
     }
 
     public void DeleteBooking(String path)
     {
+        if(path.equals("___NULL_DEV___"))
+        {
+            return;
+        }
+        final FirebaseFirestore database = FirebaseFirestore.getInstance();
+
         database.document(path)
                 .delete()
                 .addOnSuccessListener(new OnSuccessListener<Void>() {
@@ -230,7 +235,14 @@ DBookingDetails extends AppCompatActivity {
                 });
     }
 
-    public void addPatient(){
+    public void addPatient(String test){
+
+        if(test.equals("___NULL_DEV___"))
+        {
+            return;
+        }
+
+        final FirebaseFirestore database = FirebaseFirestore.getInstance();
 
         final FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
